Rename: move profile DTOs
