+++ backend/src/main/java/peer/backend/dto/profile/response/PersonalInfoResponse.java
@@ -1,4 +1,4 @@
-package peer.backend.dto.profile;
+package peer.backend.dto.profile.response;
 
 import lombok.Builder;
 import lombok.Getter;
