+++ backend/src/main/java/peer/backend/service/profile/PersonalInfoService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import peer.backend.dto.profile.PasswordRequest;
-import peer.backend.dto.profile.PersonalInfoResponse;
+import peer.backend.dto.profile.request.PasswordRequest;
+import peer.backend.dto.profile.response.PersonalInfoResponse;
 import peer.backend.entity.user.User;
 import peer.backend.exception.ForbiddenException;
 import peer.backend.exception.NotFoundException;
