+++ backend/src/main/java/peer/backend/service/profile/ProfileService.java
@@ -3,8 +3,8 @@ package peer.backend.service.profile;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import peer.backend.dto.profile.MyProfileResponse;
-import peer.backend.dto.profile.UserLinkDTO;
+import peer.backend.dto.profile.response.MyProfileResponse;
+import peer.backend.dto.profile.request.UserLinkDTO;
 import peer.backend.entity.user.User;
 import peer.backend.entity.user.UserLink;
 import peer.backend.exception.NotFoundException;
