+++ backend/src/test/java/peer/backend/profile/PersonalInfoServiceTest.java
@@ -8,8 +8,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import peer.backend.dto.profile.PasswordRequest;
-import peer.backend.dto.profile.PersonalInfoResponse;
+import peer.backend.dto.profile.request.PasswordRequest;
+import peer.backend.dto.profile.response.PersonalInfoResponse;
 import peer.backend.entity.user.User;
 import peer.backend.repository.user.UserRepository;
 import peer.backend.service.profile.PersonalInfoService;
